requirements.txt
langchain
langgraph
langchain_openai
langchain_community
openai
fastapi
uvicorn[standard]
pydantic
tqdm
httpx
scikit-learn
python-dotenv
beautifulsoup4
requests
pinecone-client
langchain-tavily
tavily-python
duckduckgo-search
lxml
voyageai
google-api-python-client
python-multipart
unstructured
pandas
